reorganize structs for service tunnel
